refactor(builder_tests): extract standing-prompt tests into separate files
Moved the large standing-prompt test functions out of `builder_tests.rs` into two new dedicated test modules, `builder_standing_prompt_tests` and `builder_standing_prompt_more_tests`, to keep the main test file focused on builder logic rather than prompt-content assertions. The extracted tests are included via `#[path]` module declarations, preserving their full coverage without bloating the original file.

Auto-committed-on: dragonfly
